Allow hiding Python interpreter name in status bar (#1)
Adds a new option `python.hideInterpreterName` (`false` by default), which, if
set to `true`, will hide the selected Python interpreter name (something like
`Python 3.8.5 64-bit ('python-3.8.5': venv)`) in the status bar. If an
interpreter was not found, the text `No Python Interpreter` will still be
displayed regardless of the setting of `python.hideInterpreterName`.